:new: Add basic struct to compilation class
